Create log_file_1
This log file is an example of log file that may be generated by a test system simulating a small space-craft launch. Each line of the log file happens to have the same number of data points, but this is only because ChatGPT was being dumb. I will try to get a better and more variable example as well.